+++ stagemonitor-spring/src/main/java/org/stagemonitor/collector/web/monitor/filter/SpringHttpExecutionContextMonitorFilter.java
@@ -8,7 +8,12 @@ import org.springframework.web.servlet.HandlerMapping;
 import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping;
 import org.stagemonitor.collector.web.monitor.SpringMonitoredHttpExecution;
 
-import javax.servlet.*;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -17,7 +22,7 @@ import java.util.Collections;
 import java.util.Enumeration;
 import java.util.List;
 
-public class SpringHttpExecutionContextMonitorFiler extends HttpExecutionContextMonitorFiler {
+public class SpringHttpExecutionContextMonitorFilter extends HttpExecutionContextMonitorFilter {
 	private final Logger logger = LoggerFactory.getLogger(getClass());
 
 	private List<HandlerMapping> allHandlerMappings;