+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # next
 ## Breaking Changes
+ * renamed HttpExecutionContextMonitorFiler to HttpExecutionContextMonitorFilter
+ * renamed SpringHttpExecutionContextMonitorFiler to SpringHttpExecutionContextMonitorFilter
 
 ## Features
  * support for monitoring ehcache (stagemonitor-ehcache)